major fix
axios replace on fetch
get data func update
fake db update
