+++ app/(components)/CardList/CardList.tsx
@@ -6,15 +6,15 @@ const { NEXT_APP_BASE_URL } = process.env;
 
 
 async function getData() {
-  const res = await axios.get(`${NEXT_APP_BASE_URL}/api/locations`);
+  const response = await fetch(`${NEXT_APP_BASE_URL}/api/locations`)
   // The return value is *not* serialized
-  // You can return Date, Map, Set, etc.
-  // Recommendation: handle errors
-  if (res.status !== 200) {
-    // This will activate the closest `error.js` Error Boundary
-    throw new Error("Failed to fetch data");
+  // You can return Date, Map, Set, etc
+  if (response.status !== 200) {
+  //   // This will activate the closest `error.js` Error Boundary
+    throw new Error('Failed to fetch data')
   }
-  return res.data;
+ 
+  return response.json();
 }
 
 export const CardList = async() => {
