+++ app/(pages)/repair/phones/page.tsx
@@ -1,18 +1,17 @@
 import React from "react";
-import axios from "axios";
 import { GadgetData, GadgetDataItem } from "@/app/(utils)/types";
 
 const { NEXT_APP_BASE_URL } = process.env;
 async function getData() {
-  const res = await axios.get(`${NEXT_APP_BASE_URL}/api/service`);
+  const response = await fetch(`${NEXT_APP_BASE_URL}/api/service`)
   // The return value is *not* serialized
-  // You can return Date, Map, Set, etc.
-  // Recommendation: handle errors
-  if (res.status !== 200) {
-    // This will activate the closest `error.js` Error Boundary
-    throw new Error("Failed to fetch data");
+  // You can return Date, Map, Set, etc
+  if (response.status !== 200) {
+  //   // This will activate the closest `error.js` Error Boundary
+    throw new Error('Failed to fetch data')
   }
-  return res.data;
+ 
+  return response.json();
 }
 export default async function Phones() {
   const data = await getData();
@@ -21,8 +20,8 @@ export default async function Phones() {
       <main>
         <h1>Послуги</h1>
         <ul>
-          {data.length
-            ? data
+          {data.length>0
+            && data
                 .filter((item:GadgetData) => item.category === "phone")[0]
                 .services.map((item:GadgetDataItem) => {
                   return (
@@ -31,7 +30,7 @@ export default async function Phones() {
                     </li>
                   );
                 })
-            : "loading"}
+            }
         </ul>
       </main>
     </>
