+++ app/(pages)/repair/powerbanks/loading.tsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+function loading() {
+  return (
+    <div>loading</div>
+  )
+}
+
+export default loading
