+++ app/api/locations/data.ts
